Added snapshot test for ChatListItem
